search cache, postmapping, putmapping

+++ src/main/java/wisemen/movienights/services/MediaService.java
@@ -4,8 +4,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 import wisemen.movienights.entities.Media;
-import wisemen.movienights.models.MediaResults;
+import wisemen.movienights.dtos.MediaResults;
+import wisemen.movienights.entities.Query;
 import wisemen.movienights.repositories.MediaRepository;
+import wisemen.movienights.repositories.QueryRepository;
 
 import java.util.List;
 import java.util.Optional;
@@ -17,14 +19,27 @@ public class MediaService {
 
     @Autowired
     private MediaRepository mediaRepository;
+    @Autowired
+    private QueryRepository queryRepository;
 
-    public List<Media> findMediaByTitle(String query) {
-        List<Media> foundMedia = restTemplate.getForObject("http://www.omdbapi.com/?apikey="+omdbKey+"&s="+query, MediaResults.class).getSearch();
-        for(int i = 0; i < foundMedia.size(); i++) {
-            foundMedia.get(i).setId(Long.parseLong(foundMedia.get(i).getImdbID().substring(2)));
-            if(!mediaRepository.existsById(foundMedia.get(i).getId())) mediaRepository.save(foundMedia.get(i));
+    public Media addMedia(Media newMedia) {
+        mediaRepository.save(newMedia);
+        return newMedia;
+    }
+
+    public List<Media> searchMediaByTitle(String query) {
+        if(queryRepository.existsById(query)) {
+            System.out.println("*from database*");
+            return mediaRepository.findMediaByTitle(query);
         }
-        return foundMedia;
+        queryRepository.save(new Query(query));
+        List<Media> mediaResults = restTemplate.getForObject("http://www.omdbapi.com/?apikey="+omdbKey+"&s="+query, MediaResults.class).getSearch();
+        for (Media media : mediaResults) {
+            media.setId(Long.parseLong(media.getImdbID().substring(2)));
+            if (!mediaRepository.existsById(media.getId())) mediaRepository.save(media);
+        }
+        System.out.println("*from omdb api*");
+        return mediaResults;
     }
 
     public Media getMediaByID(long id) {
@@ -44,11 +59,15 @@ public class MediaService {
         return foundMedia;
     }
 
-    public Boolean deleteMediaByID(long id) {
-        if(mediaRepository.existsById(id)) {
-            mediaRepository.deleteById(id);
-            return true;
-        }
-        return false;
+    public Boolean updateMediaByID(long id, Media updatedMedia) {
+        Boolean update = mediaRepository.existsById(id);
+        if(update) mediaRepository.save(updatedMedia);
+        return update;
+    }
+
+    public Media deleteMediaByID(long id) {
+        Optional<Media> databaseMedia =  mediaRepository.findById(id);
+        if(databaseMedia.isPresent()) mediaRepository.deleteById(id);
+        return databaseMedia.orElse(null);
     }
 }
